File resize option now working
It will look for the `resize` form parameter key whose value is in the format `{width}x{height}`. If this value is not set or is an empty string it will not resize
also switched from using hash-id package to short-id package. Basically because I do not have to provide the incremental value to create the hash id with short-id, which means I don't have to depend on the database to generate the autoincrement ID.
I also used this new short-id as the filename for the saved files to prevent filename collisions